Update teventcamera.cpp
- Updated update
    + Removed casting map key to size_t as it is already a size_t

+++ Scripts/src/teventcamera.cpp
@@ -182,7 +182,7 @@ void ThermalEventCamera::update(){
 	{	// update non-zero entries of output matrix with sign value
 		if(val.sign!=0)
 		{
-			this->out[(size_t)key] = val.sign;
+			this->out[key] = val.sign;
 		}
 	}
 }
